backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
bcrypt==5.0.0
